+++ src/nodejs/qm/qm_nodejs.cpp
@@ -776,7 +776,7 @@ v8::Local<v8::Value> TNodeJsStore::Field(const TQm::TRec& Rec, const int FieldId
 	} 
 	else if (Desc.IsInt64()) {
 		const int64 Val = Rec.GetFieldInt64(FieldId);
-		return HandleScope.Escape(v8::Number::New(Isolate, Val));
+		return HandleScope.Escape(v8::Number::New(Isolate, double(Val)));
 	} 
 	else if (Desc.IsByte()) {
 		const uchar Val = Rec.GetFieldByte(FieldId);
@@ -1345,7 +1345,7 @@ void TNodeJsStore::getVector(const v8::FunctionCallbackInfo<v8::Value>& Args) {
 
 			TQm::PStoreIter Iter = Store->ForwardIter(); Iter->Next();
 			for (int RecN = 0; RecN < Recs; RecN++) {
-				ColV[RecN] = JsStore->Store->GetFieldInt64(Iter->GetRecId(), FieldId);
+				ColV[RecN] = (double) JsStore->Store->GetFieldInt64(Iter->GetRecId(), FieldId);
 				Iter->Next();
 			}
 
@@ -2746,10 +2746,10 @@ void TNodeJsRecSet::filterByField(const v8::FunctionCallbackInfo<v8::Value>& Arg
         int16 MnVal = TInt16::Mn;
         int16 MxVal = TInt16::Mx;
         if (!TNodeJsUtil::IsArgNull(Args, 1) && TNodeJsUtil::IsArgFlt(Args, 1)) {
-            MnVal = TNodeJsUtil::GetArgInt32(Args, 1);
+            MnVal = (int16) TNodeJsUtil::GetArgInt32(Args, 1);
         }
         if (Args.Length() >= 3 && !TNodeJsUtil::IsArgNull(Args, 2) && TNodeJsUtil::IsArgFlt(Args, 2)) {
-            MxVal = TNodeJsUtil::GetArgInt32(Args, 2);
+            MxVal = (int16) TNodeJsUtil::GetArgInt32(Args, 2);
         }
         JsRecSet->RecSet->FilterByFieldInt16(FieldId, MnVal, MxVal);
     } else if (Desc.IsInt64()) {
